+++ pytest_cache_assert/_check_assert/caching.py
@@ -55,8 +55,12 @@ def _read_full_cache(path_cache_file: Path) -> Any:
 
 @beartype
 def write_cache_data(
-    path_cache_file: Path, *, metadata: Optional[Dict], test_data: Any, always_write: bool = False,
-) -> None:    # type: ignore[type-arg]
+    path_cache_file: Path,
+    *,
+    metadata: Optional[Dict],  # type: ignore[type-arg]
+    test_data: Any,
+    always_write: bool = False,
+) -> None:
     """Cache the specified data.
 
     Args: